Add base module, refactor
