+++ kasa/smart/modules/default_states.py
@@ -1,6 +1,9 @@
-from ..smartmodule import SmartModule
-from pydantic import BaseModel, Field
 from enum import Enum
+
+from pydantic import BaseModel, Field
+
+from ..smartmodule import SmartModule
+
 
 class Type(Enum):
     Last = "last_states"
@@ -11,25 +14,27 @@ class DefaultStatesModel(BaseModel):
     type = Field()
     state = Field()  # if type is custom, something like {"state": {"on": False}}
     """l530:
-    "state": {
-146-                "brightness": 100,
-147-                "color_temp": 9000,
-148-                "hue": 0,
-149-                "saturation": 100
-150-            },
-
+    {"state": {
+      "brightness": 100,
+      "color_temp": 9000,
+      "hue": 0,
+      "saturation": 100
+      }
+    }
     """
 
-"""L510, type stored inside brightness:
-"brightness": {
-126-                "type": "last_states",
-127-                "value": 100
-128-            },
-
-
-"""
-    re_power_type = Field()  # seen on L530, "re_power_type": "always_on", KS225 "always_off"
+    """L510, type stored inside brightness:
+    {"brightness": {
+      "type": "last_states",
+      "value": 100
+      }
+    }
+    """
+    re_power_type = (
+        Field()
+    )  # seen on L530, "re_power_type": "always_on", KS225 "always_off"
     re_power_type_capability: list | None  # KS225: ["last_states", "always_on", "always_off"]
+
 
 class DefaultStates(SmartModule):
     QUERY_GETTER_NAME = None
